+++ bench/deserializer.bench.ts
@@ -1,6 +1,6 @@
 import type { Suite } from "benchmark";
 
-import { Deserializer, Serializer } from "../../src/index.js";
+import { Deserializer, Serializer } from "../src/index.js";
 
 export default function (suite: Suite): void {
 	const data = Serializer.serialize(s => {